requirements.txt
streamlit
pytesseract
Pillow
requests
